Moved common files to components/commons folder
