+++ app/javascript/src/components/commons/EmptyState.jsx
